add: state_management_sample_container
